+++ SOPT-iOS/Projects/Data/Sources/Repository/SignInRepository.swift
@@ -36,7 +36,7 @@ extension SignInRepository: SignInRepositoryInterface {
                     case .network(let statusCode) = error,
                     statusCode == 400
                 else {
-                    return self.userService.reissuance()
+                    return self.authService.reissuance()
                 }
                 // NOTE: (@준호) 플그 미등록 + 비활동 유저의 경우 임시로 accessToken 빈 스트링 부여
                 // 자동로그인 시 활용하기 위함
