+++ SOPT-iOS/Projects/Modules/Network/Sources/Foundation/AlamoInterceptor.swift
@@ -15,7 +15,7 @@ import Alamofire
 public class AlamoInterceptor: RequestInterceptor {
     
     public typealias AdapterResult = Swift.Result<URLRequest, Error>
-    private var userService = DefaultUserService()
+    private var authService = DefaultAuthService()
 
     public func retry(_ request: Alamofire.Request, for session: Alamofire.Session, dueTo error: Swift.Error, completion: @escaping (RetryResult) -> Void) {
         // token 재발급 API가 아니며 && 로그인 실패가 아니며 && 토큰이 만료된 경우(401)
@@ -28,7 +28,7 @@ public class AlamoInterceptor: RequestInterceptor {
             return
         }
         
-        userService.reissuance { reissuanceSuccessed in
+        authService.reissuance { reissuanceSuccessed in
             if reissuanceSuccessed {
                 print("토큰 갱신 성공: ", request.request?.url)
                 completion(.retry)
