+++ SOPT-iOS/Projects/Modules/Network/Sources/Service/UserService.swift
@@ -23,8 +23,6 @@ public protocol UserService {
     func changeNickname(nickname: String) -> AnyPublisher<Int, Error>
     func getUserMainInfo() -> AnyPublisher<MainEntity, Error>
     func withdraw() -> AnyPublisher<Int, Error>
-    func reissuance(completion: @escaping ((Bool) -> Void))
-    func reissuance() -> AnyPublisher<SignInEntity, Error>
 }
 
 extension DefaultUserService: UserService {
@@ -50,32 +48,5 @@ extension DefaultUserService: UserService {
   
     public func withdraw() -> AnyPublisher<Int, Error> {
         requestObjectInCombineNoResult(.withdrawal)
-    }
-    
-    public func reissuance(completion: @escaping ((Bool) -> Void)) {
-        provider.request(.reissuance) { response in
-            switch response {
-            case .success(let value):
-                do {
-                    let decoder = JSONDecoder()
-                    let body = try decoder.decode(SignInEntity.self, from: value.data)
-                    UserDefaultKeyList.Auth.appAccessToken = body.accessToken
-                    UserDefaultKeyList.Auth.appRefreshToken = body.refreshToken
-                    UserDefaultKeyList.Auth.playgroundToken = body.playgroundToken
-                    UserDefaultKeyList.Auth.isActiveUser = body.status == .active
-                    ? true
-                    : false
-                    completion(true)
-                } catch {
-                    completion(false)
-                }
-            case .failure:
-                completion(false)
-            }
-        }
-    }
-    
-    public func reissuance() -> AnyPublisher<SignInEntity, Error> {
-        requestObjectInCombine(.reissuance)
     }
 }
